add rename playlist

+++ TestMediaApp/app/src/main/java/com/example/phamngocan/testmediaapp/function/MusicPlayer.java
@@ -44,7 +44,7 @@ public class MusicPlayer {
         context.getContentResolver().bulkInsert(uri,values );
     }
 
-    public static int createPlaylist(Context context, String name) {
+    public static long createPlaylist(Context context, String name) {
 
         if (name != null && name.length() > 0) {
             String where = MediaStore.Audio.Playlists.NAME + " = ? ";
@@ -62,7 +62,7 @@ public class MusicPlayer {
                 values.put(MediaStore.Audio.Playlists.NAME, name);
                 Uri uri = context.getContentResolver().insert(MediaStore.Audio.Playlists.EXTERNAL_CONTENT_URI,
                         values);
-                return Integer.parseInt(uri.getLastPathSegment());
+                return Long.parseLong(uri.getLastPathSegment());
 
             }
         }
