+++ TestMediaApp/app/src/main/java/com/example/phamngocan/testmediaapp/loader/PlaylistSongLoader.java
@@ -7,7 +7,6 @@ import android.provider.MediaStore;
 import com.example.phamngocan.testmediaapp.model.Song;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 
 public class PlaylistSongLoader {
     public static ArrayList<Song> getSongFromPlaylist(Context context, long playlistId) {
@@ -15,10 +14,10 @@ public class PlaylistSongLoader {
         Cursor cursor = makeCursor(context, playlistId);
         if (cursor != null && cursor.moveToFirst()) {
             do {
-                int id;
+                long id;
                 String songName,artist;
 
-                id = cursor.getInt(cursor.getColumnIndexOrThrow(MediaStore.Audio.Playlists.Members.AUDIO_ID));
+                id = cursor.getLong(cursor.getColumnIndexOrThrow(MediaStore.Audio.Playlists.Members.AUDIO_ID));
 
                 songs.add(new Song(cursor,songs.size()));
